All the scripts work

+++ extract_keys/extract_keys.sh
@@ -1,51 +1,35 @@
 #!/bin/bash
 
-$bucket_name="phase2-ceremony-files"
+bucket_name="mpc-ceremony"
+ptau_file="10.ptau"
+phase2Evaluations="phase2Evaluations"
+r1cs_file="test.r1cs"
 
-# Check the ptau file
-if [ -z "$1" ]; then
-    echo "Error: Ptau file must be provided as the first argument"
-    exit 1
-fi
-ptau_file="$1"
-if [ ! -f "$ptau_file" ]; then
-    echo "Error: The file '$ptau_file' does not exist"
-    exit 1
-fi
-
-# Check the phase2Evaluations file
-if [ -z "$2" ]; then
-    echo "Error: phase2 evaluations file must be provided as the second argument"
-    exit 1
-fi
-phase2Evaluations="$2"
-if [ ! -f "$phase2Evaluations" ]; then
-    echo "Error: The file '$phase2Evaluations' does not exist"
+# Check if Go is installed
+if ! command -v go &>/dev/null; then
+    echo "Go is not installed."
+    echo "Please install Go and try again -> https://go.dev/doc/install"
     exit 1
 fi
 
-# Check the r1cs file
-if [ -z "$3" ]; then
-    echo "Error: r1cs file must be provided as the third argument"
-    exit 1
-fi
-r1cs="$3"
-if [ ! -f "$r1cs" ]; then
-    echo "Error: The file '$r1cs' does not exist"
-    exit 1
+# Check if git is installed
+if ! git --version 2>&1 >/dev/null ; then 
+   echo "Git is not installed."
+   echo "Please install Git and try again."
+   exit 1
 fi
 
 # Check the ph2 file
 # This input can be an integer or a file path. Download if it's an integer. 
-if [ -z "$4" ]; then
-    echo "Error: Last contribution must be provided as the fourth argument"
+if [ -z "$1" ]; then
+    echo "Error: Last contribution must be provided as the first argument"
     exit 1
 fi
-latest_contribution="$4"
+latest_contribution="$1"
 if [[ $latest_contribution =~ ^[0-9]+$ ]]; then
     echo "Downloading corresponding contribution file from the bucket"
-    $latest_contribution="output_$4.ph2"
-    curl --output $latest_contribution https://$bucket_name.s3.amazonaws.com/$latest_contribution
+    latest_contribution="contribution_$1.ph2"
+    aws s3 --endpoint-url http://localhost:4566 cp s3://mpc-ceremony/$latest_contribution $latest_contribution
     echo ""
 fi
 if [ ! -f "$latest_contribution" ]; then
@@ -53,19 +37,47 @@ if [ ! -f "$latest_contribution" ]; then
     exit 1
 fi
 
+aws s3 --endpoint-url http://localhost:4566 cp s3://$bucket_name/$ptau_file $ptau_file
+aws s3 --endpoint-url http://localhost:4566 cp s3://$bucket_name/$r1cs_file $r1cs_file
+aws s3 --endpoint-url http://localhost:4566 cp s3://$bucket_name/$phase2Evaluations $phase2Evaluations
+
+# Check the ptau file
+if [ ! -f "$ptau_file" ]; then
+    echo "Error: The file '$ptau_file' does not exist"
+    exit 1
+fi
+# Check the phase2Evaluations file
+if [ ! -f "$phase2Evaluations" ]; then
+    echo "Error: The file '$phase2Evaluations' does not exist"
+    exit 1
+fi
+# Check the r1cs file
+if [ ! -f "$r1cs_file" ]; then
+    echo "Error: The file '$r1cs_file' does not exist"
+    exit 1
+fi
+
 # Check if gnark-phase2-mpc-wrapper executable exists
 # Pull and build if not so
 if [ ! -f "gnark-phase2-mpc-wrapper-executable" ]; then
+    echo ""
+    echo "Cloning and building gnark-phase2-mpc-wrapper"
     git clone https://github.com/irfanbozkurt/gnark-phase2-mpc-wrapper.git
     cd gnark-phase2-mpc-wrapper
     go mod tidy
     go build
     cd ..
-    mv gnark-phase2-mpc-wrapper/gnark-phase2-mpc-wrapper ./gnark-phase2-mpc-wrapper-executable
+    mv gnark-phase2-mpc-wrapper/gnark-phase2-mpc-wrapper gnark-phase2-mpc-wrapper-executable
     rm -rf gnark-phase2-mpc-wrapper
 fi
 
 # Extract pkey and vkey
 echo "Extracting pkey and vkey"
-./gnark-phase2-mpc-wrapper-executable extract-keys $ptau_file $latest_contribution $phase2Evaluations $r1cs
+./gnark-phase2-mpc-wrapper-executable extract-keys $ptau_file $latest_contribution $phase2Evaluations $r1cs_file
 echo ""
+
+# rm $ptau_file 
+# rm $latest_contribution
+# rm $phase2Evaluations
+# rm $r1cs_file
+# rm gnark-phase2-mpc-wrapper-executable
